cleaned up nested functions and moved to seperate files
